+++ lib/base/base_view_model.dart
@@ -2,8 +2,6 @@ import 'dart:async';
 
 import 'package:dio/dio.dart';
 import 'package:flutter/foundation.dart';
-import 'package:fpdart/fpdart.dart';
-
 import '../core.dart';
 
 export 'page_lifecycle.dart';